+++ src/index.ts
@@ -1,89 +1,33 @@
-import fs from "fs";
 import path from "path";
 import { v4 as uuidv4 } from "uuid";
 
-import DerivativesValidator from "./classes/DerivativesValidator";
 import DocumentFactory from "./classes/DocumentFactory";
-import SchemaValidator from "./classes/SchemaValidator";
-import { Config, RawConfig, Recipe } from "./interfaces/core";
-import { Derivatives } from "./interfaces/derivatives";
-import { Schema } from "./interfaces/schema";
+import { VERSION_NUMBER } from "./constants";
+import { Config } from "./interfaces/core";
 import { getTodayDate } from "./utils/dateUtils";
-import { chunkArray, createDirectory, saveAsJsonLine, zipFolder } from "./utils/fileUtils";
-
-function isValidRecipe(
-  schema?: Schema,
-  derivatives?: Derivatives,
-  reference?: Record<string, any>,
-): boolean {
-  if (schema == undefined) {
-    console.error("schema must be provided");
-    return false;
-  }
-
-  if (derivatives == undefined) {
-    console.error("derivatives must be provided, or left as an empty object");
-    return false;
-  }
-
-  if (reference == undefined) {
-    console.error("references must be provided, or left as an empty object");
-    return false;
-  }
-
-  const schemaValidator = new SchemaValidator(schema, reference);
-  const derivativesValidator = new DerivativesValidator(derivatives, schema);
-
-  schemaValidator.validateSchema();
-  derivativesValidator.validateDerivatives();
-
-  const schemaValid = schemaValidator.getValidity();
-  const derivativesValid = derivativesValidator.getValidity();
-
-  return schemaValid && derivativesValid;
-}
-
-async function saveDocuments(destDir: string, documents: object[]) {
-  createDirectory(destDir);
-
-  // each file should contain at most 10,000 items
-  const chunkedDocumentsList = chunkArray(documents, 10_000);
-  let counter = 1;
-  const filename = uuidv4();
-  for (const chunkedDocuments of chunkedDocumentsList) {
-    await saveAsJsonLine(chunkedDocuments, path.join(destDir, `${filename}-${counter}.jsonl`));
-    counter++;
-  }
-
-  // zip and delete folder afterwards
-  await zipFolder(destDir, destDir);
-  fs.rmSync(destDir, { recursive: true, force: true });
-}
-
-function loadConfig(): RawConfig {
-  const result = JSON.parse(
-    fs.readFileSync("./configs/config.json", { encoding: "utf8" }),
-  ) as RawConfig;
-
-  return result;
-}
-
-function loadRecipe(filepath: string): Recipe {
-  const result = JSON.parse(fs.readFileSync(filepath, { encoding: "utf8" })) as Recipe;
-
-  return result;
-}
+import { saveDocuments } from "./utils/fileUtils";
+import {
+  runDocumentCountPrompt,
+  runGlobalNullablePercentagePrompt,
+  runRecipeSelectionPrompt,
+  runReferenceSelectionPrompt,
+} from "./utils/prompts";
+import { isValidRecipe } from "./utils/validators/recipe-validator";
 
 async function run() {
-  const { recipeFilepath, globalNullablePercentage, documentCount, references } = loadConfig();
-  const recipe = loadRecipe(recipeFilepath);
+  console.info(`Running Forge version: ${VERSION_NUMBER}...`);
+  const recipe = await runRecipeSelectionPrompt();
+  const references = await runReferenceSelectionPrompt();
+  const documentCount = await runDocumentCountPrompt();
+  const globalNullablePercentage = await runGlobalNullablePercentagePrompt();
+
   const { schema, derivatives } = recipe;
   const destinationDir = path.join("output", getTodayDate(), uuidv4());
 
-  console.info(`Successfully retrieved recipe from: '${recipeFilepath}'. Performing validation...`);
+  console.info(`Performing recipe and reference validation...`);
 
   if (!isValidRecipe(schema, derivatives, references)) {
-    console.error("Something is wrong with the provided recipe, exiting program...");
+    console.error("Something is wrong with the provided recipe/reference, exiting program...");
     return;
   } else {
     console.info(`Validation successful! Generating ${documentCount} documents...`);